requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart
pydantic-settings
python-dotenv
google-generativeai
jinja2
loguru
alembic
requests